+++ RestaurantApi/src/main/java/com/backend/restaurantApi/model/Main.java
@@ -19,7 +19,7 @@ public class Main {
     private String dishName;
 
     @Column(name = "price")
-    private float price;
+    private double price;
 
     // used to serialize object to json
     @Override
@@ -32,12 +32,12 @@ public class Main {
                 '}';
     }
 
-    public Main(String dName, float price) {
+    public Main(String dName, double price) {
         this.dishName = dName;
         this.price = price;
     }
 
-    public Main(String dName, float price, DishAllergies da) {
+    public Main(String dName, double price, DishAllergies da) {
         this.dishName = dName;
         this.price = price;
         this.dAllergies = da;
@@ -51,7 +51,7 @@ public class Main {
         return this.id;
     }
 
-    public float getPrice() {
+    public double getPrice() {
         return this.price;
     }
 }
